backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
geoalchemy2==0.14.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.3

# Supabase
supabase>=2.3.0

# Geospatial (PostGIS handles this on DB side)
# geopandas==0.14.3  # Not needed - using PostGIS
shapely==2.0.3  # Keep for WKT parsing if needed

# Image Processing & ML
pillow==10.2.0
open-clip-torch==2.24.0
torch==2.2.0
torchvision==0.17.0

# HTTP & Storage
httpx>=0.24.0,<0.25.0
aiohttp==3.9.3
boto3==1.34.47

# Caching
redis==5.0.1
hiredis==2.3.2

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
aiofiles==23.2.1

# Error Tracking & Analytics
sentry-sdk[fastapi]==1.40.0
posthog==3.5.0